+++ react-container/src/__tests__/Container.test.tsx
@@ -1,16 +1,12 @@
+/* eslint-disable no-console */
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
 import { generateSidecar } from '__mocks/sidecar';
 
-import { mock as mockedAuthUtils } from '../__mocks/auth-utils';
 import { AuthConsumer } from '../components/AuthContainer';
 import { ContainerWithoutI18N } from '../Container';
 
-// @ts-expect-error - missing other keys
-global.console = { warn: jest.fn() };
-
-jest.mock('@cognite/auth-utils', () => mockedAuthUtils);
 jest.mock('../utils', () => {
   const utils = jest.requireActual('../utils');
   return {
@@ -21,7 +17,45 @@ jest.mock('../utils', () => {
   };
 });
 
+jest.mock('@cognite/auth-utils', () => {
+  const original = jest.requireActual('@cognite/auth-utils');
+  type callbackType = (args: unknown) => void;
+  let internalCallback: callbackType;
+  const onAuthChanged = (_applicationId: string, callback: callbackType) => {
+    internalCallback = callback;
+    return () => true;
+  };
+  return {
+    ...original,
+    getFlow: () => {
+      return { flow: 'COGNITE_AUTH' };
+    },
+    CogniteAuth: (client: unknown) => ({
+      state: {},
+      getClient: () => client,
+      onAuthChanged,
+      loginAndAuthIfNeeded: () => {
+        internalCallback({
+          client,
+          authState: { authenticated: true },
+        });
+      },
+    }),
+  };
+});
+
 describe('ContainerWithoutI18N', () => {
+  const origConsole = global.console;
+
+  beforeAll(() => {
+    // @ts-expect-error - missing other keys
+    global.console = { warn: jest.fn(), log: console.log };
+  });
+
+  afterAll(() => {
+    global.console = origConsole;
+  });
+
   it('should get correctly setup cognite client', () => {
     const Test = () => {
       const sidecar = generateSidecar();