+++ react-container/src/components/index.ts
@@ -3,3 +3,4 @@ export * from './ConditionalWrapper';
 export * from './LoopDetector';
 export * from './Sentry';
 export * from './TranslationWrapper';
+export * from './Logout';