+++ react-tenant-selector/src/components/AzureAdvancedOptions/AzureAdvancedOptions.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Input, Collapse, Button } from '@cognite/cogs.js';
+import { Input, Collapse } from '@cognite/cogs.js';
 import { useTranslation } from '@cognite/react-i18n';
 
 import { StyledAzureAdvancedOptions } from './elements';
@@ -11,15 +11,17 @@ export type AzureAdvancedOptions = {
 type Props = {
   advancedOptions: AzureAdvancedOptions;
   handleOnChange: (nextOptions: AzureAdvancedOptions) => void;
-  handleSubmit: () => void;
 };
 
-const AdvancedAzureOptions = ({
-  advancedOptions,
-  handleOnChange,
-  handleSubmit,
-}: Props) => {
+const AdvancedAzureOptions = ({ advancedOptions, handleOnChange }: Props) => {
   const { t } = useTranslation('AzureAdvancedOptions');
+
+  const handleUpdateInput = (event: React.ChangeEvent<HTMLInputElement>) => {
+    handleOnChange({
+      ...advancedOptions,
+      azureTenant: event.target.value,
+    });
+  };
 
   return (
     <StyledAzureAdvancedOptions>
@@ -28,17 +30,13 @@ const AdvancedAzureOptions = ({
           header={t('panelHeader', { defaultValue: 'Advanced Azure options' })}
         >
           <Input
-            title={t('inputTitle', { defaultValue: 'Azure tenant' })}
+            title={t('inputTitle', {
+              defaultValue: 'Azure tenant',
+            })}
             defaultValue={advancedOptions.azureTenant}
-            placeholder="Azure tenant"
-            onChange={(e) => {
-              handleOnChange({
-                ...advancedOptions,
-                azureTenant: e.target.value,
-              });
-            }}
+            placeholder="ID or name, Eg: myproject.onmicrosoft.com"
+            onChange={handleUpdateInput}
           />
-          <Button onClick={handleSubmit}>Set Azure tenant</Button>
         </Collapse.Panel>
       </Collapse>
     </StyledAzureAdvancedOptions>