+++ react-tenant-selector/src/components/LoginOptions/LoginWithAzure.spec.tsx
@@ -10,7 +10,7 @@ const getProps = () => {
   const client = new CogniteClient({ appId: 'test' });
   const authClient = new CogniteAuth(client);
 
-  authClient.login = jest.fn();
+  authClient.loginInitial = jest.fn();
 
   return { cluster: 'test', authClient };
 };
@@ -26,9 +26,28 @@ describe('<LoginWithAzure />', () => {
     render(<Base {...props} />);
 
     fireEvent.click(screen.getByText('Login with Microsoft Azure'));
-    expect(props.authClient.login).toBeCalledTimes(1);
-    expect(props.authClient.login).toHaveBeenCalledWith('AZURE_AD', {
-      cluster: 'test',
+    expect(props.authClient.loginInitial).toBeCalledTimes(1);
+    expect(props.authClient.loginInitial).toHaveBeenCalledWith({
+      flow: 'AZURE_AD',
+      directory: undefined,
+    });
+  });
+
+  it('Should trigger login on click with directory', () => {
+    const props = getProps();
+    render(<Base {...props} />);
+
+    fireEvent.click(screen.getByText('Advanced Azure options'));
+    const input = screen.getByPlaceholderText(
+      'ID or name, Eg: myproject.onmicrosoft.com'
+    );
+    fireEvent.change(input, { target: { value: 'test-directory' } });
+
+    fireEvent.click(screen.getByText('Login with Microsoft Azure'));
+    expect(props.authClient.loginInitial).toBeCalledTimes(1);
+    expect(props.authClient.loginInitial).toHaveBeenCalledWith({
+      flow: 'AZURE_AD',
+      directory: 'test-directory',
     });
   });
 });